Move fpp tests to their own package so they can run in parallel with other challenger tests.  Restart the sequencer after the test.
